worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "creative-video-worker"
version = "0.1.0"
description = "Python rendering worker for AI Creative Video Platform"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "boto3>=1.34.0",
    "Pillow>=10.2.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "moto[s3]>=4.2.0",
    "respx>=0.20.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[project.scripts]
worker = "worker.main:main"

[tool.hatch.build.targets.wheel]
packages = ["worker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
